Use config.toml file to configure SSL certificate paths
Previously, the environment var 'SSL_CERT_DIR' was used to configure
certificate because the config file was yet to be implemented then.

This change is not really backward compatible but practically, it will
just work as long as the certificate and the private key files are
stored in the same directories as recommended by the README.

A side effect of this change is that the certificate paths will be
unnecessarily stored in the app state as part of the Config struct. As
a workaround, may be all keys in the config that need to be stored in
app state can be moved under a TOML table so that only that can be
held in state.