Improve locking when reading/writing with cursor
Previously, the 'read_with_cursor' and 'write_with_cursor' functions
of 'BufferManager' used to acquire a write lock on the 'cursors'
hashmap for the entire duration of the functions. This meant that
these functions would not effectively run concurrently from multiple
threads.

In this commit, the write lock is replaced with read lock at the
beginning of the function for identifying the cursor. Once the cursor
is identified, the cursor position is maintained in local state. At
the end of the function it's written back to the cursor, while a write
lock is acquired.

Added one more set of write/read operations in tests to validate this
change.