Fix deadlock in 'BufferManager::write_with_cursor'
A write lock is being acquired on the 'region.buffer' object inside
the while loop iterations. But in the same scope,
'flush_region_if_needed' is also called, which will wait for the same
lock to be released.

A test has also been added that verifies that two threads can
concurrently write to different regions. More tests for other write
scenarios to be added in later commits.