Synchronize writes to the end of file in 'buffered_io' module
When multiple threads append i.e. write to the end of file, some
synchronization is required in the 'write_with_cursor' function. This
is because 'seek_with_cursor' and 'write_with_cursor' are two separate
calls. Hence it needs to be ensured that when multiple threads are not
writing to the end of file and updating `file_size` field
simultaneously. Otherwise threads may overwrite data written by each
other. Additionally, we also need to handle the case where `file_size`
changes between `seek_with_cursor` and `write_with_cursor` calls for
the same cursor.

This commit makes the following changes to address the above:

1. A new field 'is_eof' (bool) is added to the 'Cursor' struct. When
   'seek_with_cursor' is called, it's value will be set to true if the
   caller seeks to end of file. Otherwise it's reset to false.

2. The `file_size` field of 'BufferManager' struct is changed to use
   'RwLock' instead of 'AtomicU64'. This same field will be used to lock
   writes to end of file. (see point 3)

3. The 'write_with_cursor' function has alternate code path if it's
   writing to end of file (as indicated by 'cursor.is_eof'). In that
   case,
   a. acquire write lock on `file_size`
   b. check that cursor position = file size, if not sync it
   c. start writing in a while loop
   d. after the loop, write fize_size = curr_positon
   e. release the lock
   f. update the cursor

This commit also adds a test that verifies the scenario.