Use normal u64 instead of AtomicU64 in lru_cache module
AtomicU64 is not required here. Even though the timestamp for the
least used entry is atomically fetched, it's possible that at the time
the entry gets actually removed from the map, 'get' is called for the
same entry i.e. it will no longer be the oldest entry. Essentially,
u64 will give us the same guarantees.