Use LRUCache to cache regions in BufferManager
Earlier we were using BTreeMap, which could grow
infinitely. LRUCache suits better for this use case.

However, in this commit, we're still wrapping LRUCache in RWLock which
should be unnecessary given that the underlying datastructure is
DashMap. But the absence of RWLock causes tests to fail. To be
debugged and fixed in later commits.