Get LRUCache working in BufferManager without RwLock
Earlier, we're wrapping LRUCache in RWLock which was unnecessary given
that the underlying datastructure is DashMap. But the absence of
RWLock was causing tests to fail. This was because with the removal of
RWLock, a write lock was no longer acquired at the beginning of the
'get_or_create_region' function.

This commit fixes it by introducing a public method in LRUCache,
'get_or_insert' which can get or initialize a value without having to
externally lock the datastructure.