Use a simple counter to track age of entries in LRUCache
This replaces the earlier approach of using current timestamp. A
global counter is maintained in state. Whenever 'get', 'insert' or
'get_or_insert' methods are called, the global counter is atomically
incremented by 1 and the counter value in the corresponding row is
updated with the previous value of the global counter. This way older
entries will have smaller row counter values and can be identified as
eviction candidates.

This change also fixed the unexplained panics observed in the
tests. The previously commented out assertions are now uncommented.