Implement probabilistic eviction strategy for LRUCache
This avoids sorting of items in order to find the items that are
eligible for eviction. The tradeoff is accuracy i.e. it's possible
that an item that's evicted is newer than an older item that's not
evicted. The 'lambda' parameter can be used to tune to
'aggressiveness' of eviction.