Handle wraparound of the global counter in LRUCache
Wraparound is handled by thinking of the counter value space as a ring
buffer and then calculating 'age' of a counter associated with a
cached item depending on whether it's smaller or larger than the
global counter.

- If item counter < global counter, it means wraparound has not
  happened

- If item counter > global counter, it means wraparound has
  happened. And the 'age' of the item is calculated accordingly

Note: The above logic assumes that wraparound will happen only
once. Or in other words, after a wraparound, eviction will
successfully happen at least once until the next wraparound. This can
be improved in future by keeping track of the no. of wraparounds as
well.