Probabilistic eviction strategy for LRUCache (#89)
* Implement a method in LRUCache to iterate through values

Earlier, the LRUCache struct only provided a 'iter' method which
allowed the caller to iterate through the tuple '(V, counter)',
thereby leaking the implementation details.

This commit fixes it by providing an additional 'values' method that
returns an iterator over the values only.

* Introduce the concept of eviction strategies in LRUCache

Currently only one strategy is supported, which is 'immediate'
i.e. eviction will be attempted on every update. This is exactly how
the code behaved previously. This commit makes a provision for adding
support for multiple eviction strategies (to be done in later
commits).

* Add probablistic eviction strategy in LRUCache

Currently, this has two limitations which will be improved in later
commits.

1. The probablistic nature of this strategy is only in the sense that
   eviction of multiple items is done at a certain frequency computed
   probablistically. As we're using sorting to find the oldest keys, the
   time complexity is not optimum (O(nlogn)).This can be improved using
   probablistic approach in finding candidates for eviction as well.

2. Eviction of multiple keys is done in the same thread. It could be
   done async to improve response times.

* Trigger eviction only when an item is inserted in LRUCache

* Implement probabilistic eviction strategy for LRUCache

This avoids sorting of items in order to find the items that are
eligible for eviction. The tradeoff is accuracy i.e. it's possible
that an item that's evicted is newer than an older item that's not
evicted. The 'lambda' parameter can be used to tune to
'aggressiveness' of eviction.

* Handle wraparound of the global counter in LRUCache

Wraparound is handled by thinking of the counter value space as a ring
buffer and then calculating 'age' of a counter associated with a
cached item depending on whether it's smaller or larger than the
global counter.

- If item counter < global counter, it means wraparound has not
  happened

- If item counter > global counter, it means wraparound has
  happened. And the 'age' of the item is calculated accordingly

Note: The above logic assumes that wraparound will happen only
once. Or in other words, after a wraparound, eviction will
successfully happen at least once until the next wraparound. This can
be improved in future by keeping track of the no. of wraparounds as
well.

* Change counter type in LRUCache to AtomicU32

Now that wraparounds are handled, we can use a u32 instead of u64 for
the counter.

* Implement eviction index for LRUCache

This avoids iterating over the cache items at the time of eviction

* Fix typos pointed out by the spellchecker in github actions

* Refactor EvictionIndex in LRUCache module to use AtomicU64

.. instead of an array of Option<K> as the underlying
datastructure. This avoids the requirement of RwLock as in the
previous implementation, which would be a problem because
'on_cache_hit' method is expected to be called in hot code paths.

* Refactor ProbEviction strategy in LRUCache

The 'freq' field of the struct is now replaced by 'prob', which is
defined in terms of probability of eviction. And the 'frequency' is
now calculated as the inverse of 'prob'. The resulting behaviour is
stays the same.

* Use 'half::f16' type in ProbEviction implementation

* Fix deadlocks during probabilistic evictions in LRUCache

The 'DashMap.remove' method was being called while an immutable
reference into the map was being held (entry). This caused the
'remove' method to deadlock. See:
https://docs.rs/dashmap/latest/dashmap/struct.DashMap.html#method.remove.

* Add benchmark for lru_cache

* Move 'criterion' crate under dev-dependencies in Cargo.toml