Introduce a 'CachedValue' enum in 'LRUCache'
The 'get_or_insert' method of 'LRUCache' now returns the value wrapped
in this 'CachedValue' enum. There are 2 variants 'Hit' and 'Miss' to
indicate whether the returned value was already found in the cache or
newly inserted respectively.

This is provision for integrating the 'LRUCache' abstraction in the
'cache_loader' module (to be done in subsequent commits).