Integrating LRUCache in cache_loader module (#94)
* Fix 'get' and 'insert' methods of LRUCache

The 'index' updation code was added only to 'get_or_insert' method
previously. This commit fixes it. Also adds docstrings.

* Define an alternate constructor for LRUCache

.. as a helper to initialize it with probabilistic eviction strategy.

* Introduce a 'CachedValue' enum in 'LRUCache'

The 'get_or_insert' method of 'LRUCache' now returns the value wrapped
in this 'CachedValue' enum. There are 2 variants 'Hit' and 'Miss' to
indicate whether the returned value was already found in the cache or
newly inserted respectively.

This is provision for integrating the 'LRUCache' abstraction in the
'cache_loader' module (to be done in subsequent commits).

* Integrate 'LRUCache' in the 'cache_loader' module

The 'registry' field of 'NodeRegistry' struct is changed to use the
LRUCache abstraction.

* Make the error type generic for 'LRUCache.get_or_insert' method

This makes the 'lru_cache' module independent of the 'buffered_io'
module.