Fix 'BufferManager.seek_with_cursor' method
Earlier, the 'seek_with_cursor' method set the 'is_eof' flag for the
cursor only if the offset was 'SeekFrom::End' type. But it didn't take
into consideration that any of the 'SeekFrom' variants can result in
the current position to be at or ahead of EOF. This commit handles
such cases.