Property based tests for buffer_io module (#114)
* Add property based tests for buffer_io module

* Add tests for 'BufferManager.seek_with_cursor' method

Some failing assertions are commented out. They will be fixed in a
later commit.

* Fix 'BufferManager.seek_with_cursor' method

Earlier, the 'seek_with_cursor' method set the 'is_eof' flag for the
cursor only if the offset was 'SeekFrom::End' type. But it didn't take
into consideration that any of the 'SeekFrom' variants can result in
the current position to be at or ahead of EOF. This commit handles
such cases.

* Fix 'BufferManager.write_with_cursor' to handle writes near EOF

The 'write_with_cursor' method ends up writing past EOF in 2 cases:

1. When the caller explicitly seeks to EOF using 'SeekFrom::End(0)'
2. When the caller seeks to some location other than EOF but the sixze
   of data to be written is enough for it to write past EOF

Case 1 was handled earlier by taking a 'write' lock. But case 2 was
not handled. This commit fixes it.

* Add property based tests for BufferManager.seek_with_cursor

We just check that the method doesn't crash with u16 type for filesize
and position.