Modify 'try_get_data' method in LazyItem struct to use cache
Earlier it called the 'load_item' method of NodeRegistry which doesn't
use the LRUCache and the cuckoo filter. This commit uses 'get_object'
method which does.

To get this to work, an additional trait bound 'Cacheable' had to be
added for LazyItem<T>, because of which the same had to be specified
for InvertedIndexItem<T> (the other type besides MergedNode that
LazyItem is generic over) as well.