Implement Cacheable trait for multiple types
An earlier commit modified 'try_get_data' in LazyItem struct to use
cache. In doing so, the 'Cacheable' trait had to be added as a bound
for LazyItem method implementations. Upon rebasing the branch on
latest main, the code no longer compiled, as the following types also
required the 'Cacheable' trait to be implemented.

1. u32
2. InvertedIndexSparseAnn
3. InvertedIndexSparseAnnNode

This commit fixes it.