Refactor to initialize application context and fix lazy loading of merge nodes (#132)
* Add a doc about lmdb usage in cosdata

* Open read-only transaction to metadata db to fetch count_unindexed

Read-write transaction is not required as we're not writing anything
back.

* Introduce an 'AppContext' struct to be used as actix app state

Earlier, 'Config' was directly used as the app state. This commit
makes the provision for including more things in it.

* Handle internal errors returned by 'run_upload' function

When called from the vector API request handlers.

* Add NodeRegistry to app state (AppContext)

* Lazy load vector from disk when indexing a vector

* Add BufferManagerFactory for index files to app state

And use it to read from and write to the '{version}.index' files in
'run_upload' and 'init_vector_store' functions instead of creating a
new instance of the struct (which defeats the purpose of the caching
behaviour of BufferManager).

* Add BufferManagerFactory for vec_raw files to app state

And use it to read from and write to the '{version}.vec_raw' files.

* Modify 'try_get_data' method in LazyItem struct to use cache

Earlier it called the 'load_item' method of NodeRegistry which doesn't
use the LRUCache and the cuckoo filter. This commit uses 'get_object'
method which does.

To get this to work, an additional trait bound 'Cacheable' had to be
added for LazyItem<T>, because of which the same had to be specified
for InvertedIndexItem<T> (the other type besides MergedNode that
LazyItem is generic over) as well.

* Implement Cacheable trait for multiple types

An earlier commit modified 'try_get_data' in LazyItem struct to use
cache. In doing so, the 'Cacheable' trait had to be added as a bound
for LazyItem method implementations. Upon rebasing the branch on
latest main, the code no longer compiled, as the following types also
required the 'Cacheable' trait to be implemented.

1. u32
2. InvertedIndexSparseAnn
3. InvertedIndexSparseAnnNode

This commit fixes it.

* Get the test.py script to run completely