Persist VectorStore data in lmdb
The VectoreStore instances created upon incoming requests to the
'create-collection' endpoint were only getting added to the in-memory
DashMap and not persisted to disk.

This commit fixes that by writing the vector store data to a new lmdb
database. At startup, the data from lmdb for all collections (vector
stores) is also loaded into the DashMap. The 'VectorStoreMap' type has
been refactored to provide this abstraction.

Note that a few things are pending such as properly loading the root
node in 'VectorStoreMap' for each collection. Besides, a few more
things have been marked with TODO comments, which will be addressed in
later commits.