Add env to AppContext
So that it's stored as app state and available to the actix handler
fns. This commit also modifies all code to use the env from the app
context instead of calling the function 'get_app_env'.

There are 2 reasons for doing this change:

1. As part of initializing the 'ain_env', we're loading the vector
store map from lmdb. While doing so, we also want to eagerly load the
root vec for all vector stores. This change allows us to use the
'node_registry' (cache) in app context so that the cache is also
warmed. Note that the eager loading of root vec is pending and will be
implemented in a later commit.

2. Earlier the env was initialized lazily in the request handlers,
upon the first call to the 'get_app_env' function. But if env
initialization fails, any request will fail anwway. After this change,
the env is initialized at startup and if it fails, the process will
fail to start (code will panic). This is much better from a debugging
standpoint.