Persist VectorStore data in lmdb (#146)
* Add a doc for persisting vectore store metadata in lmdb

* Persist VectorStore data in lmdb

The VectoreStore instances created upon incoming requests to the
'create-collection' endpoint were only getting added to the in-memory
DashMap and not persisted to disk.

This commit fixes that by writing the vector store data to a new lmdb
database. At startup, the data from lmdb for all collections (vector
stores) is also loaded into the DashMap. The 'VectorStoreMap' type has
been refactored to provide this abstraction.

Note that a few things are pending such as properly loading the root
node in 'VectorStoreMap' for each collection. Besides, a few more
things have been marked with TODO comments, which will be addressed in
later commits.

* Handle WaCustomError when deleting vector store

The error is converted to 5xx (Internal server error) HTTP response.

* Add env to AppContext

So that it's stored as app state and available to the actix handler
fns. This commit also modifies all code to use the env from the app
context instead of calling the function 'get_app_env'.

There are 2 reasons for doing this change:

1. As part of initializing the 'ain_env', we're loading the vector
store map from lmdb. While doing so, we also want to eagerly load the
root vec for all vector stores. This change allows us to use the
'node_registry' (cache) in app context so that the cache is also
warmed. Note that the eager loading of root vec is pending and will be
implemented in a later commit.

2. Earlier the env was initialized lazily in the request handlers,
upon the first call to the 'get_app_env' function. But if env
initialization fails, any request will fail anwway. After this change,
the env is initialized at startup and if it fails, the process will
fail to start (code will panic). This is much better from a debugging
standpoint.

* Store offset (FileIndex) of the root node of a collection to lmdb

A new field 'file_index' is added to 'VectorStoreData' which is
serialized as CBOR and persisted to the 'collections' db of lmdb. This
will be used for locating the root vector node of the collection.

* Fix typo

* Remove the call to 'load_cache' in main fn

The load_cache call is currently failing. It will not be required,
once the cache is loaded as part of initializing the vector_store_map,
which will be implemented in subsequent commits.

* Load root nodes for all vector stores into NodeRegistry at startup

* Simplify the 'get_app_env' function

We no longer need to store the app env in a global var. Hence we can
get rid of the 'OnceLock' and have the 'get_app_env' function return
the value directly.

* Improve error handling for AppContext::new method

Earlier, the 'new' method panicked if the env could not be
initialized. This commit modifies it to return a Result with
WaCustomError. This way, the calling code may decide to panic or
handle it in some other way.

* feat(collection): use separate and only a single database for each collection

to avoid collision between versions, embeddings, branchs, and other metadata
fields, the keys are prefixed with a single byte (`u8`):

versions: 0
embeddings: 1
branchs: 2

* feat(collection): separate `prop.data`, `*.vec_raw` and `*.index` files per collection

* fix(api/transaction): use `AppContext`

---------

Co-authored-by: Vineet Naik <naikvin@gmail.com>