Use one-indexed identifiers for metadata field values
Previously, zero-indexed identifiers were used. For e.g. the values
a, b, c, d were represented as 0, 1, 2, 3. The problem with this
approach was that if the field was not specified in the input vector,
then the absence of a field value would also get represented as zero.

To fix it, the idenfiers now start with 1. So to represent x unique
values, we'd need n dimensions such that 2^n >= x + 1. This implies
that the limit on cardinality of field values is 1023 and not 1024.