Add support for metadata fields during (dense) indexing
To support metadata filtering, we are creating replica nodes at the
time of indexing (dense), where each node in a replica will have the
same prop-value but different prop-metadata (that will depend on the
fields and conditional filters to be supported for the collection as
per it's Metadata schema).

During search operation, the metadata dimensions will be appended to
the quantized vector dimensions and search will be performed with
similarly modified query vector. This part is yet to be
implemented.

Refer to 'org/metadata_filtering_hnsw.org' for feature spec.