Add code to convert from external to internal Filter type
The external 'Filter' type is what user input in json body of 'search'
API endpoint deserilizes to. The internal type is defined in the
metadata module and is used for metadata filtering implementation.

Perhaps it might be a good idea to unify these two types in
future. But for now the existing type is used as it originally defined
a more extensive API. In the previous commit, we've commented out
parts of it which are not supported by the current metadata filtering
implementation. In future, we might extend support for them.