Calculate cosine distance differently in case of indexing and query
During query, we want the metadata filters to match strongly. If
there's no strong match, whether or not the vector dimensions match
doesn't matter.

But in case of indexing, we want to match the combination of vector
dimensions and metadata dimensions (with high weight values). In this
case, there's no point in matching only the metadata dimensions.