Remove redundant set intersection operation
When computing the weighted dimensions. Earlier we were finding the
intersection of all fields supported by the schema and all fields
present in the input. However, we can assume that input fields
will be always be a subset of all fields (provided validation is taken
care of in outer layers of code).