Create replicas nodes for matching individual metadata fields
And for OR queries. Earlier we had assumed that the
all-fields-combination will be sufficient for matching individual
metadata fields as well as OR queries. But turns out that it's not
correct and results in false negatives.

To address this, we're now creating replica nodes for individual
fields too. We'd also need to create pseudo nodes for all possible
field values at the time of index creation.

Related discussion on discord -
https://discord.com/channels/1250672673856421938/1335809236537446452/1356497321537437728