Implement code to compute pseudo node dimensions
Pseudo nodes will be extra nodes that will be implicitly added to
collection indexes which support metadata filtering. The idea is that
the pseudo nodes will have random vector values but high weight values
in metadata dimensions, such that any filter condition supported by
the schema will end up matching the appropriate pseudo node. A pseudo
node with all dimensions set to high weight values will also be added
to the index such that it's connected to the root node. This way, we
can guarantee that there's a path from the root node to some node that
matches the filter condition in the query.

Related discussion on discord -
https://discord.com/channels/1250672673856421938/1335809236537446452/1356532316947615796.