Implement quantization for pseudo nodes
The pseudo nodes have all 0s in the vector embedding and we're this as
an assumption to identify if a vector is pseudo or not. As pseudo
vectors are indexed using the same code as any other input vector, we
need to make sure that the quantization step doesn't replace all 0
values. This commit handles the quantization for pseudo vector.