Modify query filter encoding
So that it works with the new implementatoin that uses pseudo nodes
and relies on comparison of mbits. This commit changes the encoding as
follows,

In case of Equal predicate, the query filter will be encoded using 0s
and 1s, just like how metadata fields are encoded during indexing.

In case of NotEqual predicate, the query filter will be encoded using
-1, 1 and 0. This is same as how it was implemented earlier.