Fix sorting of z_candidates in traverse_find_nearest fn
Earlier, it was sorting by distance (by using 'sort_by_key') and then
trying to reverse it using .rev (which wouldn't consider the
distance). This resulted in the final vector to be sorted
incorrectly. Sometimes an entry with higher distance "disappeared"
after sorting. This commit fixes it by sorting by key in the reverse
order in a single step.