Use base dimensions for root node if metadata schema is specified
Earlier the 'prop_metadata' for the root node was None even if
metadata_schema was specified for the collection. To keep things
consistent and avoid special handling, it has been fixed so that all
nodes for a collection that supports metadata filtering has metadata
dimensions (base i.e. all values equal to 0).