Optionally pass vec_id to 'traverse_find_nearest'
This is the vector id associated with the 'fvec' arg. The
'traverse_find_nearest' fn is called in case of both, querying as well
as indexing. At the time of querying, the fvec passed is query vector
which means it doesn't have any id. But during indexing, it does have
an id (specified by the user). This commit modifies the fn to accept
this id as an optional arg, so that it can be passed down to the
distance calculation functions for debugging.

Note that the vector id is the user specified id and not the derived
id obtained by calling 'ProbNode.get_id'.