Don't always create metadata dimensions for all fields combination
Earlier we were creating metadata dimensions for the all fields
combination because the older implementation relied on it for
supporting individual field queries and 'Or' queries. In the newer
implementation we are creating separate replicas to match each
individual field, which also supports 'Or' queries. Hence the all
fields combination replica that we were previously creating is no
longer required.

Note: The all fields combination replica will still be created if the
metadata schema supports 'And' condition for all fields explicitly.

Also removed some stale comments in this commit.