Calculate dims for pseudo root and other pseudo nodes separately
In changes that will be added in subsequent commits, the pseudo root
node and rest of the pseudo nodes will be created in separate code
paths (pseudo root node will be a standalone node just like the main
root node and rest of the pseudo nodes will be added under pseudo root
node). This commit is a preparation for that. For backward
compatibility, the old function is temporarily implemented using the
two new functions. It can be removed later.