Create pseudo root node separately
Previously all pseudo nodes including the pseudo root node were
created in a single 'run_upload' call. This would cause the pseudo
root node to be connected to the main root node. This commit modifies
the implementation such that the pseudo root node created separately
and it's not connected to the main root node of the index.