Reuse the reference to same NodePropValue for all metadata replicas
When the internal - external id mapping change was introduced,
indexing code was modified to create and associate a new NodePropValue
instance with every replica node in case of dense collections that
support metadata filtering. This results in multiple 'quantized_vec'
values to be held in memory.

This code fixes it by introducing a field 'replica_id' in the
'NodePropMetadata' struct which will be the actual internal_id of the
replica node. The 'ProbNode.get_id' method is modified to return this
id if 'prop_metadata' is present else fall back to
'prop_value.id' (this will be the case for all base nodes as well as
the main root and pseudo root nodes). All occurrences of direct access
to 'prob_node.prop_value.id' have been replaced with call to the
'get_id' fn (Need to enforce this but will tackle that part in a later
commit).