Get internal-external id mapping to work with metadata replica nodes
In case of collections that support metadata filtering with hnsw
index, multiple metadata replica nodes get created in the index for
every raw vector embedding. This commit gets it to work with
internal-external id mapping.

The internal ids are 'allocated' for input raw vector embeddings with
gaps based on the max no. of replica nodes supported by the index as
per the metadata schema.

Because all metadata replica nodes essentially map to the same raw
vector embedding, only the internal id of the base node is mapped to
the raw vector. When we have to obtain the raw vector for internal id,
we first calculate the base internal_id and use that to lookup the
mapping. This works because the max no. of replica nodes per vector
embedding stays constant.