Fix the 'len' and 'is_empty' methods of AtomicArray
These methods were implemented in the way that it immediately returned
the index at which a null pointer was found. This is flawed because
'push' is not the only method using which items can be added to the
array. There are other methods such as 'insert' and 'get_or_insert'
that insert an item at a specific index. So for e.g. if instead of
pushing an item into an empty array, if it's inserted at index 2, then
the length would still be 0 because the first slot (0 index) contains
a null pointer.

This commit fixes 'len' by iterating over the array and returning the
no. of slots containing non null pointers. 'is_empty' is implemented
in terms of 'len'. A basic test confirming the behavior is also added.